requirements.txt
mysql-connector-python==8.0.29
python-dotenv==0.20.0
requests==2.27.1
schedule==1.1.0

